modify id and  database

+++ src/main/java/core/dao/CoreMapper.java
@@ -9,9 +9,9 @@ public interface CoreMapper<T> {
 
     int update(T t);
 
-    int delete(int id);
+    int delete(Long id);
 
-    T getOne(int id);
+    T getOne(Long id);
 
     int getCount(T t);
